requirements.txt
beaker-pyteal==0.1.14a0
cffi==1.15.1
msgpack==1.0.4
py-algorand-sdk==1.18.0
pycparser==2.21
pycryptodomex==3.15.0
PyNaCl==1.5.0
pyteal==0.18.1
semantic-version==2.10.0
python-dotenv==0.21.0
pytest==7.1.2
docstring-parser==0.14.1
